chore: Working directory fixes

+++ src/ui/gui/dialogs/dialog_image_view/panel_histogram.cpp
@@ -14,6 +14,7 @@
 #include <qdialog.h>
 #include <qlayout.h>
 #include <qwidget.h>
+#include <cstddef>
 #include "backend/helper/image/image.hpp"
 #include "panel_image_view.hpp"
 
@@ -88,7 +89,7 @@ void PanelHistogram::drawHistogram(QPainter &painter)
     QPainterPath path;
 
     int start = 1 + mImagePanel->mutableImage()->getHistogramDisplayAreaLower();
-    path.moveTo(0, height() - hist.at<float>(start) - 2.0 * RECT_START_Y);
+    path.moveTo(0, static_cast<double>(height()) - hist.at<double>(start) - 2.0 * static_cast<double>(RECT_START_Y));
 
     for(int i = 2; i < number; i += compression) {
       int idx = i + mImagePanel->mutableImage()->getHistogramDisplayAreaLower();
@@ -102,17 +103,18 @@ void PanelHistogram::drawHistogram(QPainter &painter)
       //                  static_cast<int32_t>(startY - histValue));
 
       {
-        double x1 = (i - 1) * binWidth;
-        double y1 = height() - hist.at<float>(idx - 1) * rectHeight;
-        double x2 = i * binWidth;
-        double y2 = height() - hist.at<float>(idx) * rectHeight;
+        double x1 = static_cast<double>((i - 1)) * static_cast<double>(binWidth);
+        double y1 = static_cast<double>(height()) - hist.at<double>(idx - 1) * rectHeight;
+        double x2 = static_cast<double>(i) * static_cast<double>(binWidth);
+        double y2 = static_cast<double>(height()) - hist.at<double>(idx) * rectHeight;
 
-        double cx1 = x1 + binWidth / 3.0;
+        double cx1 = x1 + static_cast<double>(binWidth) / 3.0;
         double cy1 = y1;
-        double cx2 = x2 - binWidth / 3.0;
+        double cx2 = x2 - static_cast<double>(binWidth) / 3.0;
         double cy2 = y2;
 
-        path.cubicTo(cx1, cy1 - 2.0 * RECT_START_Y, cx2, cy2 - 2.0 * RECT_START_Y, x2, y2 - 2.0 * RECT_START_Y);
+        path.cubicTo(cx1, cy1 - 2.0 * static_cast<double>(RECT_START_Y), cx2, cy2 - 2.0 * static_cast<double>(RECT_START_Y), x2,
+                     y2 - 2.0 * static_cast<double>(RECT_START_Y));
       }
 
       //
@@ -126,7 +128,7 @@ void PanelHistogram::drawHistogram(QPainter &painter)
         painter.drawText(QRect(static_cast<int32_t>(startX - 50), static_cast<int32_t>(startY - 2 * RECT_START_Y - 4), 100, 12), Qt::AlignHCenter,
                          std::to_string(idx).data());
         painter.drawLine(static_cast<int32_t>(startX), static_cast<int32_t>(startY), static_cast<int32_t>(startX),
-                         static_cast<int32_t>(startY - static_cast<float>(rectHeight + 2 * RECT_START_Y)));
+                         static_cast<int32_t>(startY - static_cast<float>(rectHeight) + 2 * RECT_START_Y));
         painter.setPen(Qt::black);
       }
 
@@ -137,7 +139,7 @@ void PanelHistogram::drawHistogram(QPainter &painter)
         painter.drawText(QRect(static_cast<int32_t>(startX - 50), static_cast<int32_t>(startY - 2 * RECT_START_Y - 4), 100, 12), Qt::AlignHCenter,
                          std::to_string(idx).data());
         painter.drawLine(static_cast<int32_t>(startX), static_cast<int32_t>(startY), static_cast<int32_t>(startX),
-                         static_cast<int32_t>(startY - static_cast<float>(rectHeight + 2 * RECT_START_Y)));
+                         static_cast<int32_t>(startY - static_cast<float>(rectHeight) + 2 * RECT_START_Y));
         painter.setPen(Qt::black);
       }
 
@@ -148,9 +150,9 @@ void PanelHistogram::drawHistogram(QPainter &painter)
       }
     }
 
-    QColor qcolor(static_cast<int>(colors[ch][2] * 255.0f),    // R
-                  static_cast<int>(colors[ch][1] * 255.0f),    // G
-                  static_cast<int>(colors[ch][0] * 255.0f)     // B
+    QColor qcolor(static_cast<int>(colors[static_cast<size_t>(ch)][2] * 255.0F),    // R
+                  static_cast<int>(colors[static_cast<size_t>(ch)][1] * 255.0F),    // G
+                  static_cast<int>(colors[static_cast<size_t>(ch)][0] * 255.0F)     // B
     );
     QPen pen(qcolor);
     pen.setWidth(1);
